% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_if_metric_exists}
\alias{check_if_metric_exists}
\title{Check if the training (or validation) set metric exists in the multiSnpnetResults object (list)}
\usage{
check_if_metric_exists(multiSnpnetResults, metric_name)
}
\arguments{
\item{multiSnpnetResults}{a list containing the results of the multiSnpnet fit}
}
\value{
A boolean value
}
\description{
Check if the training (or validation) set metric exists in the multiSnpnetResults object (list)
}
\examples{
check_if_metric_exists(fit, 'metric_train')
check_if_metric_exists(fit, 'metric_val')
check_if_metric_exists(fit, 'AUC_train')
check_if_metric_exists(fit, 'AUC_val')

}
