% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef.multisnpnet}
\alias{coef.multisnpnet}
\title{Extract Coefficients from the Fitted Object or File}
\usage{
\method{coef}{multisnpnet}(fit = NULL, fit_path = NULL, idx = NULL, uv = TRUE)
}
\arguments{
\item{fit}{Fit object returned from multisnpnet}

\item{fit_path}{Path to the file that saves the fit object}

\item{idx}{Lambda indices where the coefficients are requested}

\item{uv}{Boolean. Whether U, V are used to represent the decomposed matrices or B, A.}
}
\value{
List of coefficients where each element is a list of one type of coefficients over the
  provided lambda indices.
}
\description{
Extract Coefficients from the Fitted Object or File
}
