% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drop_bulky_stuff_in_fit_object}
\alias{drop_bulky_stuff_in_fit_object}
\title{drop feature statistics and individual-level data}
\usage{
drop_bulky_stuff_in_fit_object(
  fit_obj,
  drop_list = c("std_obj", "response", "residuals", "stats")
)
}
\description{
drop feature statistics and individual-level data
}
