% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_all_metrics}
\alias{extract_all_metrics}
\title{Extract all weighted and unweighted metrics}
\usage{
extract_all_metrics(multiSnpnetResults)
}
\arguments{
\item{multiSnpnetResults}{a list containing the results of the multiSnpnet fit}
}
\value{
A dataframe of all available metrics
}
\description{
Extract all weighted and unweighted metrics
}
