% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Extract metrics}
\usage{
extract_metrics(multiSnpnetResults, metric_name = NULL)
}
\arguments{
\item{multiSnpnetResults}{a list containing the results of the multiSnpnet fit}

\item{metric_name}{The types of metrics (optional)}
}
\value{
A matrix of metrics (one of the followings: AUC_val, metric_val, AUC_train, and metric_train)
}
\description{
Extract metrics
}
