% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_weight}
\alias{extract_weight}
\title{Extract weights}
\usage{
extract_weight(multiSnpnetResults)
}
\arguments{
\item{multiSnpnetResults}{a list containing the results of the multiSnpnet fit}
}
\value{
A list of weights
}
\description{
Extract weights
}
