% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_prev_iter}
\alias{find_prev_iter}
\title{Get the index of the previous iteration in a specified results directory}
\usage{
find_prev_iter(results_dir, nlambda = 100)
}
\arguments{
\item{results_dir}{The results directory.}

\item{nlambda}{The maximum number of lambda}
}
\description{
Get the index of the previous iteration in a specified results directory
}
