% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cb_colors}
\alias{get_cb_colors}
\title{Generate color palette}
\usage{
get_cb_colors(key = NULL)
}
\arguments{
\item{key}{An optional argument to select a specific color in palette.}
}
\description{
Generate color palette
}
