% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_non_zero_coefficients}
\alias{get_non_zero_coefficients}
\title{Extract the non-zero coefficients from the fit object}
\usage{
get_non_zero_coefficients(fit_obj)
}
\arguments{
\item{fit_obj}{A named list containing the results of the multisnpnet results.}
}
\description{
Extract the coefficients (C) from the fit object where the coefficient has at least one non-zero entry across the response variables
}
