% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_non_zero_coefficients_as_data_frame}
\alias{get_non_zero_coefficients_as_data_frame}
\title{Extract the non-zero coefficients from the fit object and return it as a data frame}
\usage{
get_non_zero_coefficients_as_data_frame(
  multiSnpnetResults = NULL,
  fit_obj = NULL,
  genotype_file = NULL,
  withTSVD = FALSE,
  component_prefix = "Component",
  rank = NULL,
  zstdcat_path = "zstdcat"
)
}
\arguments{
\item{multiSnpnetResults}{a list containing the results of the multiSnpnet fit}

\item{fit_obj}{A named list containing the results of the multisnpnet results.}

\item{genotype_file}{Path to the new suite of genotype files. genotype_file.pvar.zst must exist.}

\item{withTSVD}{A boolean variable indicating whether to include the SVD loadings on variants}

\item{component_prefix}{(applicable when withTSVD is TRUE) the prefix of the SVD loadings column}

\item{rank}{(applicable when withTSVD is TRUE) the rank of SVD}

\item{zstdcat_path}{Path to zstdcat program, needed when loading variants.}
}
\description{
Extract the coefficients (C) from the fit object where the coefficient has at least one non-zero entry across the response variables. The function also reads the pvar file so that the resulting data frame has the chromosomal position of the genetic variants.
}
