% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_non_zero_lines}
\alias{get_non_zero_lines}
\title{Given a matrix, return lines with non-zero entries}
\usage{
get_non_zero_lines(M)
}
\arguments{
\item{M}{matrix}
}
\description{
Given a matrix, return lines with non-zero entries
}
