% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_rdata_path}
\alias{get_rdata_path}
\title{Get the path of the R data file}
\usage{
get_rdata_path(results_dir, idx)
}
\arguments{
\item{results_dir}{The results directory.}

\item{idx}{Lambda index}
}
\description{
Get the path of the R data file
}
