% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_summary_metrics}
\alias{get_summary_metrics}
\title{Tabulate the best lambda index}
\usage{
get_summary_metrics(
  multiSnpnetResults,
  metrics = c("metric_train", "metric_val", "AUC_train", "AUC_val"),
  lambda_idx = NULL
)
}
\arguments{
\item{multiSnpnetResults}{a list containing the results of the multiSnpnet fit}

\item{metrics}{a list of metrics to include}

\item{lambda_idx}{The lambda index. If not specified, we call find_best_lambda_index()}
}
\value{
A data frame of one row containing the lambda index (best and last), number of non-zero variables in the model, and the performance metrics at the best lambda index
}
\description{
For the path to a multiSnpnet results directory (saved from a multiSnpnet run with a validation set),
identify the last "previous" iteration, identify the best lambda based on the validation set AUC,
and load the lambda sequence stored in the RData file and return the lambda sequence.
This is useful for performing a "refit" using a combined set of training + validation set.
}
