% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_multiSnpnetResultsFromRDataFiles}
\alias{load_multiSnpnetResultsFromRDataFiles}
\title{Load the results of multiSnpnet}
\usage{
load_multiSnpnetResultsFromRDataFiles(
  results_dir,
  last_lambda_idx = NULL,
  fill_sequence = FALSE
)
}
\arguments{
\item{results_dir}{The results directory}

\item{last_lambda_idx}{The last lambda index. If not specified, we call find_prev_iter() and load the last lambda index available on the file system.}
}
\value{
An list object containing the training (and validation) set metrics, lambda_idx, fit, and configs. We will extract the results from the "best" lambda index for fit and configs. The configs itself is a list containing important paramters such as weight (trait weight).
}
\description{
Load the results of multiSnpnet
}
