% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisnpnet.R
\name{multisnpnet}
\alias{multisnpnet}
\title{Fit a sparse reduced rank regression model on large-scale SNP data and multivariate responses}
\usage{
multisnpnet(
  genotype_file,
  phenotype_file,
  phenotype_names,
  binary_phenotypes = NULL,
  covariate_names,
  rank,
  nlambda = 100,
  lambda.min.ratio = 0.01,
  standardize_response = TRUE,
  weight = NULL,
  split_col = NULL,
  validation = FALSE,
  mem = NULL,
  batch_size = 100,
  prev_iter = 0,
  max.iter = 10,
  configs = NULL,
  save = TRUE,
  early_stopping = FALSE
)
}
\description{
Fit a sparse reduced rank regression model on large-scale SNP data and multivariate responses
with batch variable screening and alternating minimization. It computes a full solution path
on a grid of penalty values. Can deal with larger-than-memory SNP data.
}
