% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.multisnpnet}
\alias{plot.multisnpnet}
\title{Make plots of the multisnpnet results}
\usage{
\method{plot}{multisnpnet}(
  results_dir,
  rank_prefix,
  type,
  rank,
  file_prefix,
  file_suffix,
  snpnet_dir = NULL,
  snpnet_subdir = NULL,
  snpnet_prefix = NULL,
  snpnet_suffix = NULL,
  save_dir = NULL,
  train_name = "metric_train",
  val_name = "metric_val",
  xlim = c(NA, NA),
  ylim = c(NA, NA)
)
}
\arguments{
\item{results_dir}{Character vector each specifies the parent directory of one type (e.g. exact,
lazy) of results}

\item{rank_prefix}{Character vector each specifies the prefix of the subdirectories holding
per-rank results separately}

\item{type}{Character vector each specifies the type of the results, e.g., exact, lazy}

\item{rank}{Numeric vector of ranks for which the results are available}

\item{file_prefix}{Character vector each specifies the prefix of the result file}

\item{file_suffix}{Character vector each specifies the suffix of the result file}

\item{snpnet_dir}{Parent directory the snpnet results}

\item{snpnet_subdir}{Name of the subdirectory hosting snpnet results. The result files are saved
in}

\item{save_dir}{Directory to save the plots. If NULL, no plots are generated but the list of plot
objects is returned}

\item{train_name}{Name of the object storing the validation metric}

\item{xlim}{The x limits (x1, x2) of the plots}

\item{ylim}{The y limits (y1, y2) of the plots}
}
\description{
For 50th lambda, the reduced-rank results are saved at
results_dir[i]/${rank_prefix}[i]${rank}[j]/${file_prefix}50${file_suffix}, and snpnet results are
saved in ${snpnet_dir}/${phenotype}/${snpnet_subdir}/${snpnet_prefix}50${snpnet_suffix}
}
