% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_biplot}
\alias{plot_biplot}
\title{Make biplots of the multisnpnet results}
\usage{
plot_biplot(
  svd_obj,
  component = list(x = 1, y = 2),
  label = list(phenotype = NULL, variant = NULL),
  n_labels = list(phenotype = 5, variant = 5),
  color = list(phenotype = "red", variant = "blue"),
  shape = list(phenotype = 20, variant = 4),
  axis_label = list(main = "variant", sub = "phenotype"),
  use_ggrepel = TRUE
)
}
\arguments{
\item{svd_obj}{A named list containing three matrices with u, d, and v as their names as in the
output from base::svd() function. One can pass the results of base::svd(t(fit$C)).
Please note that this function assumes svd_obj$u and svd_obj$v corresponds to phenotypes and variants, respectively.}

\item{component}{A named list that specifies the index of the components used in the plot.}

\item{label}{A named list that specifies the phenotype and variant labels.
The labels needs to be the same order as in svd_obj$u and svd_obj$v.}

\item{n_labels}{A named list that specifies the number of phenotype and variant labels in the plot.}

\item{color}{A named list that specifies the color in the plot.}

\item{shape}{A named list that specifies the color in the plot.}

\item{axis_label}{A named list that specifies the names used in the axis labels.}

\item{use_ggrepel}{A binary variable that specifies whether we should use ggrepel to annotate the
labels of the data points.}
}
\description{
Generate biplot visualization based on the decomposed coefficient matrix C.
One of the most common use case is: plot_biplot(svd(t(fit$C)), label=list('phenotype'=rownames(A_init), 'variant'=rownames(fit$C)))
}
