% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_scree}
\alias{plot_scree}
\title{Generate a scree plot visualization based on the decomposed coefficient matrix C.}
\usage{
plot_scree(svd_obj)
}
\arguments{
\item{svd_obj}{A named list containing three matrices with u, d, and v as their names as in the
output from base::svd() function. One can pass the results of base::svd(t(fit$C)) or tsvd_of_C_with_names(fit)}
}
\description{
Generate a scree plot visualization based on the decomposed coefficient matrix C.
}
\examples{
plot_scree( tsvd_of_C_with_names(fit, rank = 10) )

}
