% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_multisnpnet}
\alias{predict_multisnpnet}
\title{Predict from the Fitted Object or File}
\usage{
predict_multisnpnet(
  fit = NULL,
  saved_path = NULL,
  new_genotype_file,
  new_phenotype_file,
  idx = NULL,
  covariate_names = NULL,
  split_col = NULL,
  split_name = NULL,
  zstdcat_path = "zstdcat"
)
}
\arguments{
\item{fit}{List of fit object returned from multisnpnet.}

\item{saved_path}{Path to the file that saves the fit object. The full path is constructed as ${saved_path}${idx}.RData.}

\item{new_genotype_file}{Path to the new suite of genotype files. genotype_file.{pgen, psam, pvar.zst}.
must exist.}

\item{new_phenotype_file}{Path to the phenotype. The header must include FID, IID.}

\item{idx}{Lambda indices where the coefficients are requested.}

\item{covariate_names}{Character vector of the names of the adjustment covariates.}

\item{split_col}{Name of the split column. If NULL, all samples will be used.}

\item{split_name}{Vector of split labels where prediction is to be made.}

\item{zstdcat_path}{Path to zstdcat program, needed when loading variants}
}
\value{
A list containing the prediction and the resopnse for which the prediction is made.
}
\description{
Predict from the Fitted Object or File
}
