% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_lambda_sequence_from_multiSnpnetResults}
\alias{read_lambda_sequence_from_multiSnpnetResults}
\title{Read the lambda sequence from the initial fit with a validation set}
\usage{
read_lambda_sequence_from_multiSnpnetResults(
  multiSnpnetResults,
  lambda_idx = NULL
)
}
\arguments{
\item{lambda_idx}{The lambda index. If not specified, we call find_best_lambda_index()}

\item{results_dir}{The results directory}
}
\value{
A lambda sequence
}
\description{
For the path to a multiSnpnet results directory (saved from a multiSnpnet run with a validation set),
identify the last "previous" iteration, identify the best lambda based on the validation set AUC,
and load the lambda sequence stored in the RData file and return the lambda sequence.
This is useful for performing a "refit" using a combined set of training + validation set.
}
