% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_pvar}
\alias{read_pvar}
\title{Read the list of genetic variants present in the dataset.}
\usage{
read_pvar(genotype_file, zstdcat_path = "zstd")
}
\arguments{
\item{genotype_file}{Path to the new suite of genotype files. genotype_file.pvar.zst must exist.}

\item{zstdcat_path}{Path to zstdcat program, needed when loading variants}
}
\value{
A data frame containig the list of variants present in the genetic dataset.
}
\description{
Read the list of genetic variants present in the dataset.
}
