% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{score_contribution}
\alias{score_contribution}
\title{Compute the contribution scores}
\usage{
score_contribution(svd_obj, right_singular_vectors = FALSE)
}
\arguments{
\item{svd_obj}{A named list containing three matrices with u, d, and v as their names as in the
output from base::svd() function. One can pass the results of base::svd(t(fit$C)) or tsvd_of_C_with_names(fit).
Please note that this function assumes svd_obj$u and svd_obj$v corresponds to phenotypes and variants, respectively.}

\item{right_singular_vectors}{An indicator variable to specify if we compute the score for right singular vector or not. If true, we compute the variant squared cosine score. If false, we compute phenotype squared cosine score.}
}
\description{
Compute the relative importance traits (or variants) for each component as defined in Tanigawa et al Nat Comm 2019.
}
