% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{separate_ID_ALT_into_two_columns}
\alias{separate_ID_ALT_into_two_columns}
\title{Given a data frame with ID_ALT column, this function separates them into two columns ID and ALT}
\usage{
separate_ID_ALT_into_two_columns(ID_ALT_df)
}
\arguments{
\item{ID_ALT_df}{A dataframe with ID_ALT column}
}
\description{
Given a data frame with ID_ALT column, this function separates them into two columns ID and ALT
}
