% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tsvd_of_C_variant_loadings_as_data_framte}
\alias{tsvd_of_C_variant_loadings_as_data_framte}
\title{Extract the variant loadings of the SVD of coefficients as a data frame}
\usage{
tsvd_of_C_variant_loadings_as_data_framte(
  fit_obj,
  component_prefix = "Component",
  rank = NULL
)
}
\arguments{
\item{fit_obj}{A named list containing the results of the multisnpnet results.}

\item{component_prefix}{the prefix of the SVD loadings column}

\item{rank}{the rank of SVD}
}
\description{
Extract the variant loadings of the SVD of coefficients as a data frame
}
