% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tsvd_of_C_with_names}
\alias{tsvd_of_C_with_names}
\title{Compute the TSVD of the regression coefficient}
\usage{
tsvd_of_C_with_names(fit_obj, component_prefix = "Component", rank = NULL)
}
\arguments{
\item{fit_obj}{A named list containing the results of the multisnpnet results.}

\item{component_prefix}{A string used as a prefix for the column names corresponding to the latent variables}

\item{rank}{Desired rank of the decomposed matrices}
}
\description{
Compute the TSVD of the regression coefficient C and set colnames and rownames in the decomposed matrices.
}
