% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weighted_rowMeans}
\alias{weighted_rowMeans}
\title{weighted row means}
\usage{
weighted_rowMeans(M, w)
}
\arguments{
\item{M}{matrix}

\item{w}{weights as a named list}
}
\description{
weighted row means
}
