% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_early_stopping_condition}
\alias{check_early_stopping_condition}
\title{Check if the early stopping condition is satisfied}
\usage{
check_early_stopping_condition(ilam, metric_val, AUC_val = NULL, traits = NULL)
}
\arguments{
\item{ilam}{Lambda index}

\item{metric_val}{Validation set metric}

\item{AUC_val}{Validation set AUC}

\item{traits}{(optional) subset of traits}
}
\description{
Check if the early stopping condition is satisfied
}
