% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisnpnet.R
\name{multisnpnet}
\alias{multisnpnet}
\title{Fast Multi-Phenotype SRRR on SNP Data}
\usage{
multisnpnet(genotype_file, phenotype_file, phenotype_names, binary_phenotypes = NULL,
  covariate_names, rank, nlambda = 100, lambda.min.ratio = 0.01, standardize_response = TRUE,
  weight = NULL, validation = FALSE, split_col = NULL, mem = NULL,
  batch_size = 100, prev_iter = 0, max.iter = 10, configs = NULL, save = TRUE,
  early_stopping = FALSE)
}
\arguments{
\item{genotype_file}{Path to the suite of genotype files. genotype_file.{pgen, psam, pvar.zst}
must exist.}

\item{phenotype_file}{Path to the phenotype. The header must include FID, IID, covariate_names
and phenotype_names.}

\item{binary_phenotypes}{Names of the binary phenotypes. AUC will be evaluated for binary
phenotypes.}

\item{covariate_names}{Character vector of the names of the adjustment covariates.}

\item{rank}{Target rank of the model.}

\item{nlambda}{Number of penalty values.}

\item{lambda.min.ratio}{Ratio of the minimum penalty to the maximum penalty.}

\item{standardize_response}{Boolean. Whether to standardize the responses before fitting to deal
with potential different units of the responses.}

\item{weight}{Numberic vector that specifies the (importance) weights for the responses.}

\item{p.factor}{Named vector of separate penalty factors applied to each coefficient. This is a
number that multiplies \code{lambda} to allow different shrinkage. Default is 1 for all
variables. Can specify partially and the rest will be set to 1. Must be positive.}

\item{validation}{Boolean. Whether to evaluate on validation set.}

\item{split_col}{Name of the column in the phenotype file that specifies whether each sample
belongs to the training split or the validation split. The values are either "train" or "val".}

\item{mem}{Memory available for the program. It tells PLINK 2.0 the amount of memory it can
harness for the computation. IMPORTANT if using a job scheduler.}

\item{batch_size}{Number of variants used in batch screening.}

\item{prev_iter}{Index of the iteration to start from (e.g. to resume a previously interrupted
computation).}

\item{max.iter}{Maximum number of iterations allowed for alternating minimization.}

\item{configs}{List of additional configuration parameters. It can include:
\describe{
\item{nCores}{number of cores for the PLINK computation (default: 1)}
\item{thresh}{convergence threshold for alternating minimization (default: 1E-7)}
\item{glmnet.thresh}{convergence threshold for glmnet(Plus) (default: 1E-7)}
\item{plink2.path}{path to the PLINK2.0 program, if not on the system path}
\item{zstdcat.path}{path to the zstdcat program, if not on the system path}
\item{use_safe}{whether to use safe product to deal with very large matrix
                multiplication (default: TRUE). One may also specify MAXLEN
                (default: (2^31-1)/2), the maximum vector length passed to
                the R base matrix multiplication operation}
}}

\item{save}{Boolean. Whether to save intermediate results.}

\item{early_stopping}{Whether to stop the process early if validation metric starts to fall.}
}
\description{
Fit a sparse reduced rank regression model on large-scale SNP data and multivariate responses
with batch variable screening and alternating minimization. It computes a full solution path on a
grid of penalty values. Can deal with larger-than-memory SNP data, missing values and adjustment
covariates.
}
