% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_multiSnpnetResults}
\alias{prepare_multiSnpnetResults}
\title{Set the results of multiSnpnet run in a list}
\usage{
prepare_multiSnpnetResults(
  fit_list,
  metric_train,
  metric_val,
  AUC_train,
  AUC_val,
  configs
)
}
\arguments{
\item{configs}{}
}
\value{
An list object containing the training (and validation) set metrics, lambda_idx, fit, and configs. We will extract the results from the "best" lambda index for fit and configs. The configs itself is a list containing important paramters such as weight (trait weight).
}
\description{
Set the results of multiSnpnet run in a list
}
